Implement introduction screen
Implements an introduction scene to the start of the round after the
user has selected the difficulty. The intro scene gives a base
introduction of the context of the game and then gives a direct
interaction with the ai. The user is able to skip the intro scene if
they wish.